+++ CHANGES.md
@@ -1,5 +1,6 @@
 - Fix bug in `Message_box` color scheme handling.
   Thanks to Matthieu Dubuget for the patch (#105).
+- Add binding to `SDL_RenderGetWindow`.
 
 v1.1.0 2024-09-12 Zagreb
 ------------------------